refactor secrets-wrapper to isolate versions logic
