+++ recover/async/recombine.js
@@ -1,8 +1,8 @@
 const pull = require('pull-stream')
 const ref = require('ssb-ref')
-const secrets = require('../../secretsWrapper')
 const getContent = require('ssb-msg-content')
 
+const secrets = require('../../lib/secrets-wrapper')
 const isReply = require('../../isReply')
 
 // const pullRitual = require('../../ritual/pull/mine')
